+++ ios/StripeOnrampSdk.mm
@@ -64,6 +64,13 @@ RCT_EXPORT_METHOD(registerWalletAddress:(nonnull NSString *)address
   [StripeSdkImpl.shared registerWalletAddress:address network:network resolver:resolve rejecter:reject];
 }
 
+RCT_EXPORT_METHOD(deleteWalletAddress:(nonnull NSString *)walletId
+                              resolve:(nonnull RCTPromiseResolveBlock)resolve
+                               reject:(nonnull RCTPromiseRejectBlock)reject)
+{
+  [StripeSdkImpl.shared deleteWalletAddress:walletId resolver:resolve rejecter:reject];
+}
+
 RCT_EXPORT_METHOD(getWalletOwnershipChallenge:(nonnull NSString *)walletAddress
                                       network:(nonnull NSString *)network
                                       resolve:(nonnull RCTPromiseResolveBlock)resolve